Minimax implemented as a util
