+++ engine.py
@@ -1,0 +1,13 @@
+class ChessEngine(object):
+    def __init__(self):
+        pass
+    
+    def play(self, board):
+        raise Exception("play() function not implemented.")
+
+class MiniMaxEngine(ChessEngine):
+    def __init__(self):
+        super()
+    
+    def play(self, board):
+        pass
